Finalize entity properties

+++ src/main/kotlin/simple/marketplace/entities/Product.kt
@@ -1,5 +1,5 @@
 package simple.marketplace.entities
 
 
-class Product(val id: String, var name: String, var description: String, var price: Int) {
+class Product(val id: String, var name: String, var description: String, var price: Double) {
 }
